requirements.txt
# Framework web (compatível com Vercel)
fastapi==0.104.1
uvicorn==0.24.0

# Processamento de linguagem natural
nltk==3.8.1
textblob==0.17.1

# Machine Learning e IA
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3

# Processamento de arquivos
PyPDF2==3.0.1
python-multipart==0.0.6

# Utilitários
python-dotenv==1.0.0
requests==2.31.0

# Compatibilidade Vercel
pydantic==2.5.0
